backend/server/requirements.txt
playwright
langchain-core
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy
langchain-google-genai
python-dotenv
fastapi
uvicorn